renamed history_test and removed constructor
